Break apart monolithic EvalNode.scala
